% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Find Coefficients and P-values from Linear Model with Scatter Plot}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{a vector of outcomes.}

\item{X}{a matrix of covariates.}

\item{sub}{a list of subjects (i.e., a set of integers corresponding to rows in X).}
}
\value{

}
\description{
\code{myLinearRegression} is used to find coefficients and corresponding p-values from a linear model. Along with the outcomes, the scatter plot between each variable would be provided.
}
\examples{
## Use the data called `myData`.
## Set your working directory and read the csv file.
myData = read.csv("myData.csv")

## Dependent variable (column 1) is GPA
## Explanatory variables (column 2-6) are those:
## studying time (X1), numbers of game playing (X2),
## number of friends(X3), commute time (X4),
## and shopping time per week (X5).
simu_y = myData[, 1]
simu_x = myData[, 2:6]
simu_sub = 1:nrow(myData) # Can be any list of subjects.

## Provide coefficients and p-values of each explanatory variable.
## A scatterplot between each variable also created.
myLinearRegression(simu_y, simu_x, simu_sub)
}
